Refactor home-related entities and repositories to improve structure and organization
