+++ lib/src/data/model/product_details_model.dart
@@ -1,7 +1,7 @@
-import 'package:quickdrop/src/data/model/products_model.dart';
-import 'package:quickdrop/src/data/model/sellers_model.dart';
+import 'package:quickdrop/src/feature/home/data/model/products_model.dart';
+import 'package:quickdrop/src/feature/home/data/model/sellers_model.dart';
 import 'package:quickdrop/src/domain/entity/product_detail_entity.dart';
-import 'package:quickdrop/src/domain/entity/products_entity.dart';
+import 'package:quickdrop/src/feature/home/domain/entity/products_entity.dart';
 
 class ProductDetailsModel extends ProductDetailEntity {
   ProductDetailsModel({
