+++ lib/src/domain/repository/favorite_repository.dart
@@ -1,5 +1,5 @@
 import 'package:quickdrop/src/domain/entity/favorite_entity.dart';
-import 'package:quickdrop/src/domain/entity/products_entity.dart';
+import 'package:quickdrop/src/feature/home/domain/entity/products_entity.dart';
 
 abstract class FavoriteRepository {
   Future<FavoriteEntity> checkFavorite({
