--- lib/src/domain/repository/home_data_repository.dart
@@ -1,5 +1,0 @@
-import 'package:quickdrop/src/domain/entity/home_entity.dart';
-
-abstract class HomeDataRepository {
-  Future<HomeEntity> getHomeData();
-}
