+++ lib/src/feature/home/domain/entity/home_entity.dart
@@ -1,0 +1,18 @@
+import 'package:equatable/equatable.dart';
+import 'package:quickdrop/src/feature/home/domain/entity/products_entity.dart';
+import 'package:quickdrop/src/feature/home/domain/entity/sellers_entity.dart';
+
+class HomeEntity extends Equatable {
+  const HomeEntity({
+    required this.products,
+    required this.sellers,
+    required this.lastSeen,
+  });
+
+  final List<ProductsEntity> products;
+  final List<SellersEntity> sellers;
+  final List<ProductsEntity> lastSeen;
+
+  @override
+  List<Object> get props => <Object>[products, sellers, lastSeen];
+}
