+++ lib/src/feature/home/domain/entity/products_entity.dart
