+++ lib/src/feature/home/domain/entity/sellers_entity.dart
