+++ lib/src/feature/home/domain/repository/home_data_repository.dart
@@ -1,0 +1,5 @@
+import 'package:quickdrop/src/feature/home/domain/entity/home_entity.dart';
+
+abstract class HomeDataRepository {
+  Future<HomeEntity> getHomeData();
+}
