+++ lib/src/feature/home/domain/usecase/home_data_usecase.dart
@@ -1,0 +1,11 @@
+import 'package:quickdrop/src/feature/home/domain/entity/home_entity.dart';
+import 'package:quickdrop/src/feature/home/domain/repository/home_data_repository.dart';
+
+class HomeDataUsecase {
+  HomeDataUsecase({
+    required HomeDataRepository repository,
+  }) : _repository = repository;
+  final HomeDataRepository _repository;
+
+  Future<HomeEntity> products() async => await _repository.getHomeData();
+}
