+++ lib/src/feature/home/presentation/widgets/products/widgets/favorite/cubit/add_favorite_state.dart
