+++ lib/src/injection/injection_container.dart
@@ -3,7 +3,7 @@ import 'package:google_sign_in/google_sign_in.dart';
 import 'package:quickdrop/src/data/datasource/favorite_datasource.dart';
 import 'package:quickdrop/src/data/datasource/financial_information_datasource.dart';
 import 'package:quickdrop/src/data/datasource/firebase_auth_datasource.dart';
-import 'package:quickdrop/src/data/datasource/home_datasource.dart';
+import 'package:quickdrop/src/feature/home/data/datasource/home_datasource.dart';
 import 'package:quickdrop/src/data/datasource/my_purchases_datasource.dart';
 import 'package:quickdrop/src/data/datasource/produt_datasource.dart';
 import 'package:quickdrop/src/data/datasource/purchase_datasource.dart';
@@ -11,7 +11,7 @@ import 'package:quickdrop/src/data/datasource/purchase_detail_datasource.dart';
 import 'package:quickdrop/src/data/repository/auth_repository_impl.dart';
 import 'package:quickdrop/src/data/repository/favorite_repository_impl.dart';
 import 'package:quickdrop/src/data/repository/financial_information_repository_impl.dart';
-import 'package:quickdrop/src/data/repository/home_data_repository_impl.dart';
+import 'package:quickdrop/src/feature/home/data/repository/home_data_repository_impl.dart';
 import 'package:quickdrop/src/data/repository/my_purchases_repository_impl.dart';
 import 'package:quickdrop/src/data/repository/product_detail_repository_impl.dart';
 import 'package:quickdrop/src/data/repository/purchase_detail_repository_impl.dart';
@@ -19,7 +19,7 @@ import 'package:quickdrop/src/data/repository/purchase_repository_impl.dart';
 import 'package:quickdrop/src/domain/repository/auth_repository.dart';
 import 'package:quickdrop/src/domain/repository/favorite_repository.dart';
 import 'package:quickdrop/src/domain/repository/financial_information_repository.dart';
-import 'package:quickdrop/src/domain/repository/home_data_repository.dart';
+import 'package:quickdrop/src/feature/home/domain/repository/home_data_repository.dart';
 import 'package:quickdrop/src/domain/repository/my_purchases_repository.dart';
 import 'package:quickdrop/src/domain/repository/product_detail_repository.dart';
 import 'package:quickdrop/src/domain/repository/purchase_detail_repository.dart';
@@ -27,7 +27,7 @@ import 'package:quickdrop/src/domain/repository/purchase_repository.dart';
 import 'package:quickdrop/src/domain/usecase/auth_usecase.dart';
 import 'package:quickdrop/src/domain/usecase/favorite_usecase.dart';
 import 'package:quickdrop/src/domain/usecase/financial_information_usecase.dart';
-import 'package:quickdrop/src/domain/usecase/home_data_usecase.dart';
+import 'package:quickdrop/src/feature/home/domain/usecase/home_data_usecase.dart';
 import 'package:quickdrop/src/domain/usecase/my_purchases_usecase.dart';
 import 'package:quickdrop/src/domain/usecase/produc_detail_usecase.dart';
 import 'package:quickdrop/src/domain/usecase/purchase_detail_usecase.dart';
@@ -35,8 +35,8 @@ import 'package:quickdrop/src/domain/usecase/purshase_usecase.dart';
 import 'package:quickdrop/src/presentation/app/cubit/app_cubit.dart';
 import 'package:quickdrop/src/presentation/favorites/cubit/favorites_cubit.dart';
 import 'package:quickdrop/src/presentation/financial_information/cubit/financial_information_cubit.dart';
-import 'package:quickdrop/src/presentation/home/cubit/home_cubit.dart';
-import 'package:quickdrop/src/presentation/home/widgets/products/widgets/favorite/cubit/add_favorite_cubit.dart';
+import 'package:quickdrop/src/feature/home/presentation/cubit/home_cubit.dart';
+import 'package:quickdrop/src/feature/home/presentation/widgets/products/widgets/favorite/cubit/add_favorite_cubit.dart';
 import 'package:quickdrop/src/presentation/login/cubit/login_cubit.dart';
 import 'package:quickdrop/src/presentation/my_purchases/cubit/my_purchases_cubit.dart';
 import 'package:quickdrop/src/presentation/product/productCubit/product_cubit.dart';
