+++ lib/src/presentation/product/widgets/relatedProducts/related_product.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:quickdrop/src/core/constants/constants.dart';
 import 'package:quickdrop/src/core/functions/price_formatter.dart';
-import 'package:quickdrop/src/domain/entity/products_entity.dart';
+import 'package:quickdrop/src/feature/home/domain/entity/products_entity.dart';
 import 'package:quickdrop/src/presentation/widgets/clipper_radius_images.dart';
 import 'package:quickdrop/src/presentation/widgets/image_loader.dart';
